### 2025-08-15 #8820 @editorialbot
[PRE REVIEW]: STEPS: Serially Transferred Evolving Population Simulator
**Submitting author:** @Devinmlake (<a href="http://orcid.org/0000-0002-9974-2367">Devin Lake</a>)
**Repository:** https://github.com/zachmatson/STEPS.git
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.0

## Editor notes

Invites: @bricoletc @jonaspleyer @Ladme @helmutcarter
