### 2025-12-14 c8d6d9a
Remove slow use of fmod
