### 2026-01-28 6db0986
Updated JOSS Manuscript
Updated the paper.md file to include new sections and added a reference to paper.bib
