### 2026-02-03 #8820 @ethanwhite
Hi @Devinmlake - I'll be handling your paper. Sorry it's taken us a while to get to it and that this means you're caught in the middle of our scope and format transition. Please take a look at the [message about these changes above](https://github.com/openjournals/joss-reviews/issues/8820#issuecomment-3765336358). Once you've made the appropriate changes let me know and I'll do a check on whether the software meets the new scope requirements and if it does move forward with the review process.
