### 2026-02-04 a6db89b
Documentation update (#6)
* Create STEPS_User_Manual.pdf

Uploading the v2.0.0 user manual to repository ahead of the code update.

* JOSS Submission Files

Added a folder containing the markdown file for the JOSS manuscript as well as the .bib file and the 3 figures.

* Update README.md

* Updated JOSS Manuscript

Updated the paper.md file to include new sections and added a reference to paper.bib
