+++ JOSS/paper.bib
@@ -292,4 +292,18 @@ abstract = {Advances in modern biology have allowed us to measure evolutionary f
 	title = {{STEPS} User Manual (v2.0.0)},
 	url = {https://doi.org/10.5281/zenodo.16690374}}
 	
+@article{Smith:2016,
+	author = {Smith, James J. and Johnson, Wendy R. and Lark, Amy M. and Mead, Louise S. and Wiser, Michael J. and Pennock, Robert T.},
+	year = {2016},
+	url = {https://doi.org/10.1186/s12052-016-0060-0},
+	journal = {Evolution: Education and Outreach},
+	title = {An Avida-ED digital evolution curriculum for undergraduate biology},
+	doi = {10.1186/s12052-016-0060-0},
+	volume = {9},
+	number = {1},
+	pages = {9},
+	isbn {1936-6434},
+	id = Smith2016}
+	}
+	
 	
