### 2026-02-23 #8820 @ethanwhite
@Devinmlake There are a couple of bibtex typos preventing me from rendering the pdf for review. PR submitted to clean them up: https://github.com/zachmatson/STEPS/pull/7
