### 2026-02-25 #8820 @danielskatz
@ethanwhite - when you start a scope evaluation, it's good practice to explain why, both for the authors and for other editors
