### 2026-02-25 #8820 @ethanwhite
Thanks @danielskatz - do you typically do that in the issue or over on the scope review page
