### 2026-02-25 #8820 @danielskatz
I do it in the issue, so that the author understands what we are questioning, and perhaps can weigh in if they have something to add.
