### 2026-03-09 #8820 @ethanwhite
Thanks @danielskatz!

@Devinmlake - apologies for not communicating about the scope review process (this is my first time doing one). As mentioned above we're in the midst of a transition at JOSS in the criteria used to determine whether or not a particular software project is in-scope (driven in part by a surge in LLM-based submissions). Given this transition I wasn't sure if your submission quite met some of the new development history/approach requirements. Fortunately there is now general agreement that you are in scope so we can proceed!

During the scope review there were a couple of things mentioned that will definitely get flagged by reviewers since they are on our review checklist. I'm going to start looking for reviewers, but I'd suggest that while that process is happening you go ahead and make a couple of changes:
1. Add a clear pathway for community contributions (e.g., a [CONTRIBUTING.md file](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) that includes information on how to a) Contribute to the software; b) Report issues or problems with the software; and c) Seek support.
2. Add some automated tests (preferably running automatically using GitHub Actions)

Finding reviewers can take some time, so if you can take care of these and any other missing items from [the checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) while that is happening it will facilitate a faster review once things get started.
