### 2026-03-09 #8820 @ethanwhite
👋 @bricoletc @jonaspleyer @Ladme @helmutcarter - would you be willing to review this submission for JOSE? I recognize that the topic fit may not be perfect for some of you, but all of you have experience with Rust and simulation and adding in evolution puts too strong a filter on our reviewer pool. We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://openjournals.readthedocs.io/en/jose/review_criteria.html. Let me know if you have any questions
