### 2026-03-17 5732391
Mutator mutations
Added a new kind of mutation that allows the lineages mutation rates to increase by 100-fold or decrease back to the ancestral rate.
