+++ .github/workflows/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,21 @@
+<!--
+Thank you for contributing to this project! You must fill out the information below before we can review this pull request. By explaining why you're making a change (or linking to an issue) and what changes you've made, we can triage your pull request to the best possible team for review.
+-->
+
+### Why:
+
+<!-- Paste the issue link or number here -->
+Closes: 
+
+<!-- If there's an existing issue for your change, please link to it above.
+If there's _not_ an existing issue, please open one first to make it more likely that this update will be accepted: https://github.com/zachmatson/STEPS/issues/new. -->
+
+### What's being changed (if available, include any code snippets, screenshots, or gifs):
+
+<!-- Let us know what you are changing. Share anything that could provide the most context. -->
+
+### Check off the following:
+
+- [ ] A subject matter expert (SME) has reviewed the technical accuracy of the content in this PR. In most cases, the author can be the SME.
+- [ ] The changes in this PR meet [the docs fundamentals that are required for all content](https://github.com/zachmatson/STEPS?tab=CONTRIBUTING).
+- [ ] All CI checks are passing and the changes look good in the review environment.
