### 2026-03-23 1676d7f
JOSS Formatting
Fixed the formatting for the pdf generation and created a spot for the statement of need section
