+++ JOSS/paper.md
@@ -32,38 +32,48 @@ affiliations:
   index: 4
 - name: DeSC Healthcare, Shibuya, Tokyo, Japan 
   index: 5
-date: 13 August 2025
+date: 28 January 2026
 bibliography: paper.bib
 ---
-# Summary
 
+# Summary
 Bacteria and other microbes are widely used in the field of experimental evolution [@Lenski:2017; @McDonald:2019; @Ascensao:2025]. In many of these experiments, the populations are propagated by periodic serial dilutions into fresh media, with the microbes growing to a maximum population size between successive transfers. Here we present the Serially Transferred Evolving Population Simulator (STEPS) software, which simulates the dynamics of asexual populations as they grow and evolve in these experiments. The STEPS software allows new mutations to occur while a population grows, including beneficial, neutral, and deleterious mutations. The resulting lineages then grow at different rates according to the fitness effects of their accumulated mutations. After the total population reaches the maximum size, the dilution factor determines the proportion of individuals that are randomly chosen to start the next growth cycle. The differential growth rates of the lineages reflect the process of natural selection, while the periodic bottlenecks impose random genetic drift.
 
-The underlying mechanics of the STEPS program are designed to simulate the biological processes that occur in actual microbial evolution experiments, while taking advantage of computational methods that allow the simulations to proceed quickly and efficiently. For example, the timesteps used in the simulations coincide roughly with the population’s doubling time, providing meaningful timepoints for introducing mutations. The mutational categories, the effects of mutations on relative growth rates, and the serial transfer process have been implemented such that the dynamics of the simulated populations reflect the same interplay of mutation, selection, and drift that occurs in biological experiments.
-
 # Statement of need
+The Long-Term Evolution Experiment (LTEE) with *Escherichia coli* was started in 1988 [@Lenski:1991; @Lenski:2023] and it continues to this day [@Barrick:2023]. The LTEE has been used to study many aspects of evolutionary dynamics [@Blount:2008; @Wiser:2013; @Tenaillon:2016; @Couce:2024], and it has inspired many other experiments that use similar approaches [@Johnson:2021; @Izutsu:2024; @Stroud:2025]. Various publications based on these projects have included simulations to analyze and interpret the experimental results. The authors of these papers have written new simulation software that is custom-built to a specific experimental system or result, and even publications based on the same system may use different simulation methods [e.g., @Good:2017; @Wiser:2013]. We anticipate that future studies will be able to use STEPS to perform relevant quantitative analyses or interpret qualitative results. Researchers may also use STEPS to design new experiments or as a framework on which to build customized simulations. We expect that employing the STEPS program in these various ways will improve consistency and efficiency across studies.
 
-The Long-Term Evolution Experiment (LTEE) with *E. coli* was started in 1988 [@Lenski:1991; @Lenski:2023] and it continues to this day [@Barrick:2023]. The LTEE has been used to study many aspects of evolutionary dynamics [@Blount:2008; @Wiser:2013; @Tenaillon:2016; @Couce:2024], and it has inspired many other experiments that use similar approaches [@Johnson:2021; @Izutsu:2024; @Stroud:2025]. Various publications based on these projects include simulations to analyze and interpret the experimental results. The authors of these papers typically write new simulation software that is custom-built to a specific experimental system or result, and even publications based on the same system may use different simulation methods [e.g., @Good:2017; @Wiser:2013]. We anticipate that future studies will be able to use STEPS to perform relevant quantitative analyses or interpret qualitative results. Researchers may also use STEPS to design new experiments or as a framework for building customized simulations. We expect that employing the STEPS program in these various ways will improve consistency and efficiency across studies.
+# State of the field
[…]
