+++ JOSS/paper.md
@@ -42,6 +42,9 @@ Bacteria and other microbes are widely used in the field of experimental evoluti
 # Statement of need
 The Long-Term Evolution Experiment (LTEE) with *Escherichia coli* was started in 1988 [@Lenski:1991; @Lenski:2023] and it continues to this day [@Barrick:2023]. The LTEE has been used to study many aspects of evolutionary dynamics [@Blount:2008; @Wiser:2013; @Tenaillon:2016; @Couce:2024], and it has inspired many other experiments that use similar approaches [@Johnson:2021; @Izutsu:2024; @Stroud:2025]. Various publications based on these projects have included simulations to analyze and interpret the experimental results. The authors of these papers have written new simulation software that is custom-built to a specific experimental system or result, and even publications based on the same system may use different simulation methods [e.g., @Good:2017; @Wiser:2013]. We anticipate that future studies will be able to use STEPS to perform relevant quantitative analyses or interpret qualitative results. Researchers may also use STEPS to design new experiments or as a framework on which to build customized simulations. We expect that employing the STEPS program in these various ways will improve consistency and efficiency across studies.
 
+# State of the field
+There are a few existing tools that are commonly used for simulating large evolving populations. Two of the most popular are Avida [@Smith:2016], SLiM [@]. 
+
 # Software Design
 To enable STEPS to be accessible to as many researchers as possible, we have produced two different ways to run the simulations. The first approach requires using a command-line interface to run STEPS, and it is intended for advanced applications, such as generating data for large numbers of replicate populations for statistical analysis and comparison to experimental results. The second approach uses a web-based graphical user interface that provides immediate visualization of the results, helping researchers develop their intuition about the dynamics of evolving populations under various scenarios. This version of STEPS can also be used in educational settings, allowing students without computational backgrounds to obtain results quickly and easily. The web-based version was inspired, in part, by the Avida-ED software used to promote a better understanding of evolution through experiments with agent-based digital organisms in a virtual world [@Smith:2016]. Some features of the command-line version of STEPS are not available in the web-based version. However, both versions use the same underlying code while the simulations are running.
 
@@ -50,7 +53,7 @@ The underlying mechanics of the STEPS program are designed to simulate the biolo
 # Research impact statement
 The first version of STEPS was developed for and used extensively in a peer-reviewed paper on the effects of population bottlenecks on the dynamics of experimentally evolving bacteria [@Izutsu:2024]. Since then, the STEPS program has undergone significant development including (i) refinements to speed up the simulations, (ii) creation of a web-based version to complement the command-line version, and (iii) writing an in-depth user manual to help both beginning and advanced users better understand both the STEPS software and the population dynamical processes that it simulates. The user manual is freely available [@Lake:2025] and has been downloaded more than 250 times since it was posted (August 1, 2025).
 
- With the rapid growth of both experimental evolution and microbial genomics, we expect many research groups will find similar utility in the STEPS program, and we look forward to its further development by ourselves and others for future needs that may arise. Indeed, we are using STEPS in our ongoing research. For example, it is difficult to estimate the fraction of observed mutations that are beneficial drivers versus neutral hitchhikers using empirical data alone. However, one can estimate that fraction by comparing empirical data with STEPS simulations that include only beneficial mutations with alternative distributions that also include neutral and even deleterious mutations.
[…]
