### 2026-03-25 #10300 @editorialbot
[REVIEW]: STEPS: Serially Transferred Evolving Population Simulator
**Submitting author:** @Devinmlake (<a href="http://orcid.org/0000-0002-9974-2367">Devin Lake</a>)
**Repository:** https://github.com/zachmatson/STEPS.git
**Branch with paper.md** (empty if default branch): main
**Version:** v2.0.0
