### 2026-04-01 #10300 @bricoletc
@Devinmlake @zachmatson there's a typo [on this line](https://github.com/zachmatson/STEPS/blob/6dc2ba186a53206ec66726101846e2093f24bb0a/JOSS/paper.md?plain=1#L60) preventing Figure 1 to appear in the paper PDF (JOSS -> Joss in linking to Figure 1 file name I believe)
