### 2026-04-02 3cb9f1d
JOSS Manuscript
A "State of the field" section has been added to the manuscript along with the necessary updates to the .bib file for references.
