+++ JOSS/paper.bib
@@ -306,4 +306,75 @@ abstract = {Advances in modern biology have allowed us to measure evolutionary f
 	id = Smith2016}
 	}
 	
-	
+@book{Garland:2009,
+	title = {Experimental evolution: concepts, methods, and applications of selection experiments},
+	author = {Garland, Theodore and Rose, Michael R},
+	year = {2009},
+	publisher = {Univ of California Press}
+	}
+
+@book{Kassen:2024,
+	title = {Experimental evolution and the nature of biodiversity},
+	author = {Kassen, Rees},
+	year = {2024},
+	publisher = {Oxford University Press}
+	}
+
+@book{Rose:2026,
+	title = {A Primer for Experimental Evolution},
+	author = {Rose, Michael R and Lesnki, Richard E and Matos, M and Graves, Joseph L},
+	year = {2026},
+	publisher = {World Scientific Publishing, Singapore}
+	}
+
+@incollection{Deathrage:2014,
+	title = {Identification of mutations in laboratory-evolved microbes from next-generation sequencing data using breseq},
+	author = {Deatherage, Daniel E and Barrick, Jeffrey E},
+	booktitle = {Engineering and analyzing multicellular systems: methods and protocols},
+	pages = {165--188},
+	year = {2014},
+	publisher = {Springer}
+	}
+
+@article{Zibbu:2025,
+	title = {seabreeze: A Pipeline for Analyzing Structural Variation Between Bacterial Genome Assemblies},
+	author = {Zibbu, Ira and Wilke, Claus O and Barrick, Jeffrey E},
+	journal = {Journal of Open Source Software},
+	volume = {10},
+	number = {111},
+	year = {2025},
+	publisher = {Journal of Open Source Software}
+	}
+
+@article{Ofria:2004,
+	title = {Avida: A software platform for research in computational evolutionary biology},
+	author = {Ofria, Charles and Wilke, Claus O},
+	journal = {Artificial life},
+	volume = {10},
+	number = {2},
+	pages = {191--229},
+	year = {2004},
+	publisher = {MIT Press}
+	}
+
+@article{Covert:2013,
+	title = {Experiments on the role of deleterious mutations as stepping stones in adaptive evolution},
+	author = {Covert III, Arthur W and Lenski, Richard E and Wilke, Claus O and Ofria, Charles},
+	journal = {Proceedings of the National Academy of Sciences},
+	volume = {110},
+	number = {34},
+	pages = {E3171--E3178},
+	year = {2013},
+	publisher = {National Academy of Sciences}
+	}
+
+@article{Haller:2026,
+	title = {SLiM 5: Eco-evolutionary simulations across multiple chromosomes and full genomes},
+	author = {Haller, Benjamin C and Ralph, Peter L and Messer, Philipp W},
+	journal = {Molecular Biology and Evolution},
+	volume = {43},
+	number = {1},
+	pages = {msaf313},
+	year = {2026},
+	publisher = {Oxford University Press UK}
+}
