+++ JOSS/paper.md
@@ -43,9 +43,9 @@ Bacteria and other microbes are widely used in the field of experimental evoluti
 The Long-Term Evolution Experiment (LTEE) with *Escherichia coli* was started in 1988 [@Lenski:1991; @Lenski:2023] and it continues to this day [@Barrick:2023]. The LTEE has been used to study many aspects of evolutionary dynamics [@Blount:2008; @Wiser:2013; @Tenaillon:2016; @Couce:2024], and it has inspired many other experiments that use similar approaches [@Johnson:2021; @Izutsu:2024; @Stroud:2025]. Various publications based on these projects have included simulations to analyze and interpret the experimental results. The authors of these papers have written new simulation software that is custom-built to a specific experimental system or result, and even publications based on the same system may use different simulation methods [e.g., @Good:2017; @Wiser:2013]. We anticipate that future studies will be able to use STEPS to perform relevant quantitative analyses or interpret qualitative results. Researchers may also use STEPS to design new experiments or as a framework on which to build customized simulations. We expect that employing the STEPS program in these various ways will improve consistency and efficiency across studies.
 
 # State of the field
-There are a few existing tools that are commonly used for simulating large evolving populations. Two of the most popular are Avida [@Smith:2016], SLiM [@]. 
+Experimental evolution is a vibrant field of research that has expanded greatly over the past few decades [@Garland:2009; @Kassen:2024; @Rose:2026]. This expansion has been driven by the recognition that asexual microbial populations provide powerful systems for observing evolutionary dynamics, and by the increasing feasibility of analyzing complete genomes [@Deatherage:2014; @Zibbu:2025]. As noted above, some previous simulation tools have been developed for these kinds of experimental systems, but on an ad hoc basis and customized for a specific project. Asexual agent-based models also exist (e.g., [@Ofria:2004; @Covert:2013]), but they are not intended to simulate actual biological organisms, and they run slowly owing to the execution of the programs of the individual agents. Finally, there are simulation programs for evolving populations of sexual organisms (e.g., [@Haller:2026]), but these generally assume initial genetic variation and recombination that do not occur in most evolution experiments with microbes. Thus, there is great value in having an accessible and flexible simulation platform like STEPS to advance the field.
 
-# Software Design
+# Software design
 To enable STEPS to be accessible to as many researchers as possible, we have produced two different ways to run the simulations. The first approach requires using a command-line interface to run STEPS, and it is intended for advanced applications, such as generating data for large numbers of replicate populations for statistical analysis and comparison to experimental results. The second approach uses a web-based graphical user interface that provides immediate visualization of the results, helping researchers develop their intuition about the dynamics of evolving populations under various scenarios. This version of STEPS can also be used in educational settings, allowing students without computational backgrounds to obtain results quickly and easily. The web-based version was inspired, in part, by the Avida-ED software used to promote a better understanding of evolution through experiments with agent-based digital organisms in a virtual world [@Smith:2016]. Some features of the command-line version of STEPS are not available in the web-based version. However, both versions use the same underlying code while the simulations are running.
 
 The underlying mechanics of the STEPS program are designed to simulate the biological processes that occur in actual evolution experiments with microorganisms, while taking advantage of computational methods that allow the simulations to proceed quickly and efficiently. For example, the timesteps in the simulations correspond approximately to the population’s doubling time, providing meaningful timepoints for introducing mutations. The mutational categories, the effects of mutations on growth rates, and the serial transfer process have been implemented such that the dynamics of the simulated populations reflect the same interplay of mutation, selection, and drift that occurs in biological experiments. The underlying theoretical framework for STEPS was inspired, in part, by simulations published previously [@Good:2017; @Wiser:2013]. However, STEPS has been designed and written to allow for easier access, wider use, and future development.
