### 2026-04-02 #10300 @Devinmlake
@bricoletc @jonaspleyer I've made some edits to the manuscript including a fix to the pdf generation and added the State of the field section. Those changes are on the "Documentation-Update" branch for now. I will try to keep any updates related to the JOSS review on that branch until things are closer to finished and ready to be merged with the main branch.
