### 2026-04-04 #10300 @jonaspleyer
> [@bricoletc](https://github.com/bricoletc) [@jonaspleyer](https://github.com/jonaspleyer) I've made some edits to the manuscript including a fix to the pdf generation and added the State of the field section. Those changes are on the "Documentation-Update" branch for now. I will try to keep any updates related to the JOSS review on that branch until things are closer to finished and ready to be merged with the main branch.

Thanks for letting us know. However to be able to compile the new document via the `editorialbot`, you will have to push it to the default branch. Otherwise we might have to change the branch specified in the submission.
