### 2026-04-06 #10300 @ethanwhite
Thanks @jonaspleyer & @Devinmlake. Devin - as Jonas notes it would be nice to fix this particular issue fixed in `main` now so that the reviewers can easily view the whole manuscript with all of the figures. If that's an issue for some reason let me know.
