### 2026-04-13 #10300 @ethanwhite
Looks like I can get us a rendered version of the pdf with the fixes in @Devinmlake's branch through the editorial bot, so I'm going to try to do that now to help everyone more forward with their reviews.
