+++ JOSS/paper.bib
@@ -302,7 +302,7 @@ abstract = {Advances in modern biology have allowed us to measure evolutionary f
 	volume = {9},
 	number = {1},
 	pages = {9},
-	isbn {1936-6434},
+	isbn = {1936-6434},
 	id = Smith2016}
 	}
 	
