### 2026-04-13 #10300 @ethanwhite
I've submitted a PR with the fix needed to render the updated PDF https://github.com/zachmatson/STEPS/pull/9
