### 2026-04-20 #10300 @ethanwhite
@Devinmlake - just checking in our getting your branch fixed or merging those changes into main so that we can render the current version of the paper for the reviewers
