### 2026-04-20 fc77242
Adding the GitHub Action to compile the JOSS manuscript
* Create draft-pdf.yml

Adding the GitHub Action to compile the JOSS manuscript

* Update draft-pdf.yml

Fixed the path to the markdown file for the paper
