### 2026-04-21 #10300 @rlenski
@ethanwhite I've sent you an email with some very sad news. Coauthor Zach Matson is working to fix the remaining issues with the pdf.
