### 2026-04-21 #10300 @rlenski
@zachmatson  Figure 1 is still missing from the proofs. The file name in text must match exactly (including JOSS vs Joss) the file name in the folder that contains the figure.
