### 2026-04-21 5bff82c
Fix paper rendering issues
