+++ JOSS/paper.bib
@@ -327,7 +327,7 @@ abstract = {Advances in modern biology have allowed us to measure evolutionary f
 	publisher = {World Scientific Publishing, Singapore}
 	}
 
-@incollection{Deathrage:2014,
+@incollection{Deatherage:2014,
 	title = {Identification of mutations in laboratory-evolved microbes from next-generation sequencing data using breseq},
 	author = {Deatherage, Daniel E and Barrick, Jeffrey E},
 	booktitle = {Engineering and analyzing multicellular systems: methods and protocols},
