+++ .github/workflows/draft-pdf.yml
@@ -21,4 +21,5 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: paper.pdf
+          path: JOSS/paper.pdf
+          if-no-files-found: error
