### 2026-04-21 #10300 @ethanwhite
Thank you @zachmatson and @rlenski for getting the paper rendering.

@jonaspleyer & @bricoletc - you should be all set to proceed now. For context, I think it's important to let you both know that the [lead author passed away unexpectedly](https://telliamedrevisited.wordpress.com/2026/04/09/a-profound-loss-in-the-ltee-family/) and so the other authors will be handling the responses. It is part of the JOSS culture to be kind, collaborative, and supportive during the review process. In this case it will be especially important for us to embody this approach.
