### 2026-04-21 #10300 @jonaspleyer
I am left devastated and speechless by these news. In science we sometimes overlook that it is not only about publications, grants and conferences. But it is also about the journey and the people that we meet along the way. And in this very moment I am reminded of this fact in the most saddening way possible.

My deepest condolences and best wishes go out to his family, colleagues and friends.

Rest in peace Devin Lake
