### 2026-04-22 #10300 @bricoletc
Dear authors, here is my review below.

First of all, like many in the field of evolution, I am a big admirer of the Long-Term Evolution Experiment (LTEE)!
I am clearly convinced of the value of the STEPS software, whose design should enable recovering, modelling and fitting
parameters to the LTEE observations. More broadly, it applies to any experimental evolution setups involving
asexually reproducing microbial populations. The authors have also gone to great lengths - with a highly detailed
user manual and interactive and user-friendly web-based tool - to make STEPS accessible to a wider audience
beyond researchers. This is also highly valuable as understanding evolution is not trivial, but highly important.

My comments are below. I have just one major comment relating to explicitly emphasising
the extent to which STEPS-based simulations fit LTEE observations.
My other comments are minor and relate to clarity and documentation.

I would also like to extend my deepest sympathies regarding the passing away of the lead author
of this paper, and I hope this review and paper will help honour their memory.

## Paper
- Realised impact: You motivate well how software design fits biological experiments (L72-75)
    and that STEPS was first developed to model/understand the effect of bottlenecks in evolving bacteria (L79-81).
    To add to this, please state explicitly to what extent you observe good fits between STEPS-based
    simulations and LTEE observations - in Izutsu et al. (2024) and perhaps other studies.
    You could add this in either of the lines I mentioned above.
- L45: what is an agent-based model? A model where each individual is explicitly modelled? This
    is probably worth stating explicitly.
    Also, what limitation are you referring to when you say 'they are not intended to simulate
    actual biological organisms'?
- In Figure 1, the notion of 'Average' is not defined.
    Is it an average across all lineages in one population at a given time point? And also a rolling average
    accounting for localised population bottlenecks? Please provide an explicit definition.
    This will avoid thinking it is an average across individuals in a population, and reinforce
    your earlier statements that you do not explicitly model individuals.
- Figure 1: What is 'accumulated mutations' exactly?. Is it sum of all fixed mutations? And here, presumably,
    given the simulation parameters, fixed beneficial mutations only?
    Being more explicit here, in the figure legend or the y-axis title, would help.
- Figure 1 legend: what is strength of epistasis? Either define it, or remove it from the legend for simplicity.

## Tool use
- Usability: I tried the command-line (CLI) and web-server versions of STEPs, both worked great!
    I have a question regarding CLI however, from the output TSV, is there an easy or recommended
    way to plot the results? I.e., to produce your Figure 1/the figures in the web-based version.
    If so, please indicate in the paper; if not, please consider providing a means to do this.

## Tool tests
- It is not clear if STEPS has tests (functional, integration tests), or how to run them.
    If none, you could provide at least a small simulation example command with expected output documented in the README,
    so that users can verify proper running.

## Tool documentation
- Github README: please put the compilation and running commands under '```sh ```' markdown blocks,
    this will more obviously flag them as blocks to run in a shell prompt.
- Github README (2): please provide a small section outlining how others can contribute, report issues, seek support.
