### 2026-04-22 #10300 @rlenski
Thank you, Brice, for this helpful and constructive review, and for your condolences on Devin's passing.  We will endeavor to address the points you raise. It might take a little more time than usual, as I'm traveling over the next two weeks.  I will also need to learn how to use the GitHub and JOSS tools to edit the manuscript and files as needed, while working closely with Zach Matson on the technical issues as well.
