### 2026-05-20 zachmatson/STEPS#13 @jonaspleyer
[JOSS] How to test web-based software locally?
As by the JOSS submission guidelines
https://joss.readthedocs.io/en/latest/submitting.html
web-based software must be testable locally. I could not find how to do this.

https://github.com/openjournals/joss-reviews/issues/10300#issuecomment-4176156639
