### 2026-05-20 zachmatson/STEPS#17 @jonaspleyer
[JOSS] Missing Community Guidelines
I could not find any community guidelines of how to submit changes/code to this repository.

https://github.com/openjournals/joss-reviews/issues/10300#issuecomment-4176156639
