### 2026-08-30 zachmatson/STEPS#17 @zachmatson
Closed by merging #28. Thank you for raising this @goelakash.

@jonaspleyer let me know if the changes address your concern.
